test/regression/reftest/web-platform-tests/dom/events/Event-dispatch-reenter.html

<!DOCTYPE html>
<meta charset=utf-8>
<!--EDIT-->
<!--title> Dispatch additional events inside an event listener </title-->
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<style>
  #table {
    display: none;
  }
  .col {
    display: inline-block;
  }
  .cell {
    border: 1px solid black;
    padding: 0;
    margin: 1px;
  }
</style>
<div id=log></div>
<div id="table">
  <div id="table-body">
    <div class="col" id="parent">
      <p class="cell">Shady Grove</p>
      <p class="cell" id="target">Over the river, Charlie</p>
    </div>
    <div class="col">
      <p class="cell">Aeolian</p>
      <p class="cell">Dorian</p>
    </div>
  </div>
</div>
<script>
test(function() {
    var event_type = "foo";
    var target = document.getElementById("target");
    var parent = document.getElementById("parent");
    var tbody = document.getElementById("table-body");
    var table = document.getElementById("table");
    var body = document.body;
    var html = document.documentElement;
    var targets = [window, document, html, body, table, tbody, parent, target];
    var expected_targets = [
        window, document, html, body, table,
        target, parent, tbody,
        table, body, html, document, window,
        tbody, parent, target];
    var actual_targets = [];
    var expected_types = [
        "foo", "foo", "foo", "foo", "foo",
        "bar", "bar", "bar",
        "bar", "bar", "bar", "bar", "bar",
        "foo", "foo", "foo"
    ];

    var actual_targets = [], actual_types = [];
    var test_event = this.step_func(function(evt) {
        actual_targets.push(evt.currentTarget);
        actual_types.push(evt.type);
        if (table == evt.currentTarget && event_type == evt.type) {
            // EDIT
            // var e = document.createEvent("Event");
            // e.initEvent("bar", true, true);
            var e = new Event("bar", {"bubbles":true, "cancelable":true});
            target.dispatchEvent(e);
        }
    });

    for (var i = 0; i < targets.length; ++i) {
        targets[i].addEventListener(event_type, test_event, true);
        targets[i].addEventListener("bar", test_event, false);
    }

    // EDIT
    // var evt = document.createEvent("Event");
    // evt.initEvent(event_type, false, true);
    var evt = new Event(event_type, {"bubbles":false, "cancelable":true});
    target.dispatchEvent(evt);

    assert_array_equals(actual_targets, expected_targets, "actual_targets");
    assert_array_equals(actual_types, expected_types, "actual_types");
});
</script>
